Moved Menu component to new Menu directory
